mlai/Langchain/requirements.txt
# LangChain Core
langchain
langchain-core

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

--extra-index-url https://download.pytorch.org/whl/cu124
torch
